requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.16
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-simple-history==3.3.0
django-timezone-field==5.1
django_advance_thumbnail==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.2
h11==0.14.0
httptools==0.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.3.1
MarkupSafe==2.1.5
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.23.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
